Uploading of Tomcat files successfully

+++ deployer/receiver.py
@@ -11,19 +11,30 @@ from subprocess import Popen
 import subprocess
 import pwd
 
+import sys
+sys.path.append(os.path.realpath(__file__))
+import conf
+
 class DeployHandler(RequestHandler):
 	@tornado.web.asynchronous
 	def post(self):
 		file1 = self.request.files['file'][0]
 		
 		command = self.get_argument('command', '')
-		folder = self.get_argument('folder', '')
+		idpolo = self.get_argument('idpolo', '')
+		tomcat = self.get_argument('tomcat', False)
+		
+		if not tomcat:
+			folder = self.get_argument('folder', '')
+		else:
+			folder = conf.TOMCAT_PATH
 		fname = file1['filename']
+		
 		user = self.get_argument('user', '')
 		
 		user_pwd = pwd.getpwnam(user)
 		
-		if folder == "":
+		if folder == '':
 			folder = user_pwd.pw_dir
 				
 		#deployment_folder = os.path.join(user_home, folder)
@@ -35,7 +46,7 @@ class DeployHandler(RequestHandler):
 
 		@tornado.gen.coroutine
 		def call_execute():
-			yield thread_pool.submit(self.execute, command=command, file_desc=file1, filename=final_directory, directory=folder, user=user_pwd)
+			yield thread_pool.submit(self.execute, command=command, file_desc=file1, filename=final_directory, directory=folder, user=user_pwd, tomcat=tomcat)
 			
 		thread_pool = futures.ThreadPoolExecutor(max_workers=1)
 		tornado.gen.Task(call_execute)
@@ -43,16 +54,21 @@ class DeployHandler(RequestHandler):
 		self.finish('OK')
 
 	@tornado.web.asynchronous
-	def execute(self, command, file_desc, filename, directory, user):
+	def execute(self, command, file_desc, filename, directory, user, tomcat=False):
 		
 		def demote(user_uid, user_gid):
 			os.setgid(user_gid)
 			os.setuid(user_uid)
 
-		output_file = open(filename, 'wb')
+		if os.path.exists(os.dirname(filename)):
+			output_file = open(filename, 'wb')
+		else:
+			return
 		output_file.write(file_desc['body'])
-		os.chown(filename, user.pw_uid, user.pw_gid)
-		
+		if not tomcat:
+			os.chown(filename, user.pw_uid, user.pw_gid)
+		else:
+			os.chown(filename, pwd.getpwnam('tomcat7').pw_uid, pwd.getpwnam('tomcat7').pw_gid)
 		output_file.close()
 		if command is not "":
 			process = Popen(split(command), preexec_fn=demote(user.pw_uid, user.pw_gid), cwd=directory, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
@@ -89,5 +105,5 @@ if __name__ == "__main__":
 		"cert_reqs": ssl.CERT_OPTIONAL,
 		"ca_certs": conf.APPCERT,
 		})
-	httpServer.listen(1339)
+	httpServer.listen(conf.RECEIVER_PORT)
 	ioloop.IOLoop.instance().start()
